Added notification frequency tests
