+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/store/HealthStoreInterface.kt
@@ -1,0 +1,11 @@
+package com.outsystems.plugins.healthfitnesslib.store
+
+import com.outsystems.plugins.healthfitnesslib.HealthFitnessError
+
+interface HealthStoreInterface {
+
+    fun getVariableByName(name : String) : GoogleFitVariable?
+    fun advancedQueryAsync(parameters : AdvancedQueryParameters,
+                           onSuccess : (AdvancedQueryResponse) -> Unit,
+                           onError : (HealthFitnessError) -> Unit)
+}
