+++ tests/android/HealthFitnessTests/app/build.gradle
@@ -31,10 +31,9 @@ android {
         unitTests.includeAndroidResources = true
     }
 }
-
 dependencies {
 
-    implementation fileTree(dir: "libs", include: ["*.jar"])
+
     implementation("androidx.appcompat:appcompat:1.3.1")
     implementation("androidx.test:core-ktx:1.4.0")
     implementation("com.github.OutSystems.os-plugins-base-interface:oscore-android:core-0.0.3@aar")
@@ -48,7 +47,8 @@ dependencies {
     testImplementation("org.jetbrains.kotlin:kotlin-test")
     testImplementation("androidx.test.ext:junit-ktx:1.1.3")
     testImplementation("com.github.OutSystems.os-plugins-base-interface:oscore-android:core-0.0.3@aar")
-    testImplementation("com.github.OutSystems.os-plugins-base-interface:oshealthfitness-android:hf-0.0.23@aar")
+    //testImplementation("com.github.OutSystems.os-plugins-base-interface:oshealthfitness-android:hf-0.0.23@aar")
+    testImplementation fileTree(dir: "libs", include: ["*.jar", "*.aar"])
 
     androidTestImplementation("androidx.test.ext:junit:1.1.3")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.4.0")
